requirements.txt
streamlit>=1.24.0
openpyxl>=3.0.10
pandas>=1.5.0
python-dotenv>=0.21.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.3
langchain>=0.0.200
langchain-community>=0.0.10
langchain-groq>=0.1.0
groq>=0.3.0
langchain-huggingface>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
transformers
nltk